update Klader and user pages
