+++ public/locales/en/translation.json
@@ -33,7 +33,9 @@
     "edit": "Edit",
     "delete": "Delete",
     "addNewUser": "New Users",
-    "addNew": "Add New"
+    "newkladers": "New Kladers",
+    "editKladers": "Add New",
+    "addKladers": "Add kladers"
   },
   "form": {
     "addNewUser": "Add New User",
