+++ src/pages/add_new_User/view_user.scss
@@ -31,10 +31,7 @@
     }
   }
 
-  .general_information_section,
-  .address_section,
-  .formations_section,
-  .experiences_section {
+  .general_information_section {
     box-shadow: 2px 4px 10px 1px rgba(201, 201, 201, 0.47);
     display: grid;
     grid-template-columns: 1fr; /* Default to a single column */
@@ -76,149 +73,6 @@
       opacity: 50%;
     }
   }
-  .address_details {
-    display: grid;
-    grid-template-columns: 1fr ; /* Default to a single column */
-    gap: 15px;
-    padding: 0 2rem 2rem 2rem;
-    .input-group {
-      display: flex;
-      flex-direction: row;
-      gap: 10px;
-
-      label {
-        font-weight: bold;
-      }
-      div {
-        border-radius: 5px;
-        margin-bottom: 0.5rem;
-      }
-    }
-
-    .horizontal_line {
-      color: #ccc;
-      opacity: 50%;
-    }
-  }
-  .address_details {
-    grid-template-columns: 1fr 1fr;
-  }
-  .formations_details,
-  .experiences_details {
-    grid-template-columns: 1fr;
-    padding: 0 2rem 2rem 2rem;
-    .icon_wrapper {
-      display: flex;
-      align-items: center;
-      gap: 20px;
-      margin-bottom: 1rem;
-      .icon {
-        width: 50px;
-        height: 50px;
-        border-radius: 50%;
-        background-color: #cccccc53;
-        display: inline-flex;
-        justify-content: center;
-        align-items: center;
-      }
-
-      .school_wrapper {
-        .school_heading {
-          font-weight: bold;
-          font-size: 18px;
-        }
-        .school_place {
-          font-size: 14px;
-          font-weight: 400;
-        }
-      }
-    }
-    .school_year {
-      font-weight: bold;
-      font-size: 16px;
-      span {
-        font-size: 14px;
-        font-weight: 400;
-      }
-    }
-    .designation_year {
-      font-weight: bold;
-      font-size: 16px;
-      span {
-        font-size: 14px;
-        font-weight: 400;
-      }
-    }
-    .activity-wrapper {
-      font-weight: bold;
-      font-size: 16px;
-      .activity-content {
-        display: flex;
-        gap: 20px;
-        font-weight: 100;
-        margin-top: 1rem;
-        flex-wrap: wrap;
-
-        .activity-box {
-          border-radius: 30px;
-          padding: 0.5rem;
-          border: 1px solid #ccc;
-          font-size: 14px;
-          font-weight: 100;
-        }
-      }
-    }
-    .skills_wrapper {
-      font-weight: bold;
-      font-size: 16px;
-      margin-top: 0.5rem;
-
-      .skills-content {
-        display: flex;
-        gap: 20px;
-        font-weight: 100;
-        margin-top: 1rem;
-        flex-wrap: wrap;
-
-        .skills-box {
-          border-radius: 30px;
-          padding: 0.5rem;
-          border: 1px solid #ccc;
-          font-size: 14px;
-          font-weight: 100;
-        }
-      }
-    }
-    .experience_wrapper {
-      font-weight: bold;
-      font-size: 16px;
-      margin-top: 0.5rem;
-
-      .experience_place {
-        font-weight: 100;
-      }
-
-      .designation-content {
-        display: flex;
-        gap: 20px;
-        font-weight: 100;
-        margin-top: 1rem;
-        flex-wrap: wrap;
-
-        .designation-box {
-          border-radius: 30px;
-          padding: 0.5rem;
-          border: 1px solid #ccc;
-          font-size: 14px;
-          font-weight: 100;
-        }
-      }
-    }
-    .experience_designation {
-      font-weight: 600;
-      margin-top: 0.5rem;
-    }
-  }
 
   @media (min-width: 768px) {
     .general_information_details {
@@ -239,125 +93,6 @@
         gap: 1rem;
       }
     }
-
-    .address_details {
-      grid-template-columns: 1fr 1fr 1fr;
-    }
-    .formations_details,
-    .experiences_details {
-      grid-template-columns: 1fr;
-      .icon_wrapper {
-        display: flex;
-        align-items: center;
-        gap: 20px;
-        margin-bottom: 1rem;
-        .icon {
-          width: 50px;
-          height: 50px;
-          border-radius: 50%;
-          background-color: #cccccc53;
-          display: inline-flex;
-          justify-content: center;
-          align-items: center;
-        }
-
-        .school_wrapper {
-          .school_heading {
-            font-weight: bold;
-            font-size: 18px;
-          }
-          .school_place {
-            font-size: 14px;
-            font-weight: 400;
-          }
-        }
-      }
-      .school_year {
-        font-weight: bold;
-        font-size: 16px;
-        span {
-          font-size: 14px;
-          font-weight: 400;
-        }
-      }
-      .designation_year {
-        font-weight: bold;
-        font-size: 16px;
-        span {
-          font-size: 14px;
-          font-weight: 400;
-        }
-      }
-      .activity-wrapper {
-        font-weight: bold;
-        font-size: 16px;
-        .activity-content {
-          display: flex;
-          gap: 20px;
-          font-weight: 100;
-          margin-top: 1rem;
-          flex-wrap: wrap;
-
-          .activity-box {
-            border-radius: 30px;
-            padding: 0.5rem;
-            border: 1px solid #ccc;
-            font-size: 14px;
-            font-weight: 100;
-          }
-        }
-      }
-      .skills_wrapper {
-        font-weight: bold;
-        font-size: 16px;
-        margin-top: 0.5rem;
-
-        .skills-content {
-          display: flex;
-          gap: 20px;
-          font-weight: 100;
-          margin-top: 1rem;
-          flex-wrap: wrap;
-
-          .skills-box {
-            border-radius: 30px;
-            padding: 0.5rem;
-            border: 1px solid #ccc;
-            font-size: 14px;
-            font-weight: 100;
-          }
-        }
-      }
-      .experience_wrapper {
-        font-weight: bold;
-        font-size: 16px;
-        margin-top: 0.5rem;
-
-        .experience_place {
-          font-weight: 100;
-        }
-
-        .designation-content {
-          display: flex;
-          gap: 20px;
-          font-weight: 100;
-          margin-top: 1rem;
-          flex-wrap: wrap;
-
-          .designation-box {
-            border-radius: 30px;
-            padding: 0.5rem;
-            border: 1px solid #ccc;
-            font-size: 14px;
-            font-weight: 100;
-          }
-        }
-      }
-      .experience_designation {
-        font-weight: 600;
-        margin-top: 0.5rem;
-      }
-    }
   }
 
   @media (min-width: 1024px) {
@@ -365,14 +100,6 @@
       .right {
         grid-template-columns: repeat(3, 1fr);
       }
-    }
-
-    .address_details {
-      grid-template-columns: 1fr 1fr 1fr 1fr; /* Default to a single column */
-    }
-    .formations_details,
-    .experiences_details {
-      grid-template-columns: 1fr;
     }
   }
 }
