+++ src/pages/list/KladersListing.jsx
@@ -1,0 +1,13 @@
+import KladersDatatable from "../../components/datatable/KladersDatatable";
+import "./kladersListing.scss";
+import React from "react";
+
+const KladersListing  = () => {
+  return (
+    <>
+      <KladersDatatable />
+    </>
+  );
+};
+
+export default KladersListing;
