--- src/pages/list/List.jsx
@@ -1,13 +1,0 @@
-import "./list.scss";
-import Datatable from "../../components/datatable/Datatable";
-import React from "react";
-
-const List = () => {
-  return (
-    <>
-      <Datatable />
-    </>
-  );
-};
-
-export default List;
