+++ src/pages/list/UserListing.jsx
@@ -1,0 +1,14 @@
+import UserDatatable from "../../components/datatable/UserDatatable";
+import "./userListing.scss";
+import React from "react";
+
+
+const UserListing = () => {
+  return (
+    <>
+      <UserDatatable />
+    </>
+  );
+};
+
+export default UserListing;
