+++ src/pages/list/userListing.scss
@@ -1,0 +1,7 @@
+.list{
+    display:flex;
+    width:100%;
+    .listContainer{
+       flex:6;
+    }
+}
